feat: 후원자 - 펀딩 모금하기 API(카카오페이 제외)  구현

+++ src/main/java/gdg/whowantit/dto/fundingDto/FundingResponseDto.java
@@ -1,11 +1,7 @@
 package gdg.whowantit.dto.fundingDto;
 
 import gdg.whowantit.entity.ApprovalStatus;
-import gdg.whowantit.entity.Beneficiary;
 import gdg.whowantit.entity.Status;
-import jakarta.persistence.Column;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -70,7 +66,7 @@ public class FundingResponseDto {
 
         private Long fundingRelationId;
 
-        private Long paymentAmount;
+        private float paymentAmount;
 
     }
